requirements.txt
-i https://pypi.org/simple
aldryn-apphooks-config==0.5.2
aldryn-boilerplates==0.8.0
aldryn-common==1.0.5
aldryn-search==1.0.1
backports.csv==1.0.7
boto3==1.9.122
botocore==1.12.122
certifi==2019.3.9
chardet==3.0.4
cmsplugin-filer==1.1.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
django-acme-challenge==1.2
django-appconf==1.0.3
django-appdata==0.2.1
django-autocomplete-light==3.3.2
django-classy-tags==0.8.0
django-cms==3.6.0
django-cors-headers==2.5.2
django-daterange-filter==1.3.0
django-filer==1.4.4
django-filter==2.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mass-edit==3.2.0
django-meta-mixin==0.3.0
django-meta==1.5.0
django-mptt==0.9.1
django-multiselectfield==0.1.8
django-object-actions==1.0.0
django-parler==1.9.2
django-paypal==1.0.0
django-polymorphic==2.0.3
django-resized==0.3.9
django-rest-swagger==2.2.0
django-sekizai==0.10.0
django-simple-history==2.7.0
django-sortedm2m==1.5.0
django-spurl==0.6.5
django-standard-form==1.1.1
django-storages==1.7.1
django-taggit-autosuggest==0.3.4
django-taggit-templatetags==0.2.5
django-taggit==0.23.0
django-templatetag-sugar==1.0
django-treebeard==4.3
django-webpack-loader==0.6.0
django==1.11.20 ; python_version >= '3'
djangocms-admin-style==1.3.0
djangocms-apphook-setup==0.3.0
djangocms-attributes-field==1.0.0
djangocms-blog==0.9.9
djangocms-file==2.2.0
djangocms-inherit==0.2.2
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-teaser==0.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework==3.9.2
docutils==0.14
easy-thumbnails==2.6
et-xmlfile==1.0.1
futures==3.1.1
googlemaps==3.0.2
gunicorn==19.6.0
html5lib==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
jinja2==2.10
jmespath==0.9.4
lxml==4.3.3
markdown==3.1
markupsafe==1.1.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.1
pillow==5.4.1 ; python_version >= '2.7'
psycopg2-binary==2.7.7
python-dateutil==2.8.0
pytz==2018.9
pyyaml==5.1
requests==2.21.0
rollbar==0.14.7
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
tablib==0.13.0
unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1 ; python_version >= '3.4'
urlobject==2.4.3
webencodings==0.5.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
yurl==0.13
